fix(serve): the /auth route still need to contain "Authorization" header issue
- add the /auth route checking in auth credential middleware.
- refactor integration-testing example structure and add authenticate example
