+++ packages/integration-testing/src/example1/projectConfig.ts
@@ -1,20 +1,15 @@
-import {
-  VulcanBuilder,
-  IBuildOptions,
-  SchemaReaderType,
-} from '@vulcan-sql/build';
+import { SchemaReaderType } from '@vulcan-sql/build';
 import {
   ArtifactBuilderProviderType,
   ArtifactBuilderSerializerType,
   TemplateProviderType,
   DocumentSpec,
 } from '@vulcan-sql/core';
-import { VulcanServer, ServeConfig, APIProviderType } from '@vulcan-sql/serve';
+import { APIProviderType } from '@vulcan-sql/serve';
 import * as path from 'path';
-import * as supertest from 'supertest';
 import faker from '@faker-js/faker';
 
-const projectConfig: ServeConfig & IBuildOptions = {
+export default {
   name: 'example project 1',
   description: 'Vulcan project for integration testing',
   version: '0.0.1',
@@ -53,35 +48,3 @@ const projectConfig: ServeConfig & IBuildOptions = {
   },
   port: faker.datatype.number({ min: 20000, max: 30000 }),
 };
-
-let server: VulcanServer;
-
-afterEach(async () => {
-  await server?.close();
-});
-
-it.each([
-  [
-    '436193eb-f686-4105-ad7b-b5945276c14a',
-    [
-      {
-        id: '436193eb-f686-4105-ad7b-b5945276c14a',
-        name: 'ivan',
-      },
-    ],
-  ],
-  ['2dc839e0-0f65-4dba-ac38-4eaf023d0008', []],
-])(
-  'Example1: Build and serve should work',
-  async (userId, expected) => {
-    const builder = new VulcanBuilder(projectConfig);
-    await builder.build();
-    server = new VulcanServer(projectConfig);
-    const httpServer = (await server.start())['http'];
-
-    const agent = supertest(httpServer);
-    const result = await agent.get(`/api/user/${userId}`);
-    expect(JSON.stringify(result.body)).toEqual(JSON.stringify(expected));
-  },
-  10000
-);
