+++ packages/serve/src/lib/middleware/auth/authCredentialMiddleware.ts
@@ -57,6 +57,9 @@ export class AuthCredentialMiddleware extends BuiltInMiddleware<AuthOptions> {
 
     const options = this.getOptions() as AuthOptions;
 
+    // The /auth endpoint not need contains "Authorization" in header and auth credentials
+    if (context.path === '/auth') return;
+
     // pass current context to auth token for users
     for (const name of Object.keys(this.authenticators)) {
       // skip the disappeared auth type name in options
